moved main.py to root folder
